Inline PayloadValuesGenerator
