+++ testing/features/support/drivers/all.rb
@@ -6,4 +6,3 @@ require_relative "browserstack/android"
 require_relative "browserstack/chrome"
 require_relative "browserstack/internet_explorer"
 require_relative "browserstack/ios"
-require_relative "browserstack/payload_values_generator"
