+++ testing/features/support/helpers/env_variables.rb
@@ -109,5 +109,9 @@ module Helpers
     def log_location
       ENV["LOG_LOCATION"]
     end
+
+    def sha
+      @sha ||= `git rev-parse --short HEAD`.chomp
+    end
   end
 end
